Hack in support of "partially" random data
This commit changes the random button in the gui to no longer actually
randomly shuffle the data and instead to shuffle the data and then sort
the first 90% of it. As a result, the first 90% of the data are already
in sorted order and then the last 10% are random.

The purpose of this is to add in support for data generation on which
timsort is designed to perform well so that it can be compared to
mergesort to make the difference between them more readily apparent.

Support for this was just hacked in because wx-programming is apparently
beyond my mental capabilities. If this were of general interest, it
should be added in as its own drop down option instead of replacing the
Random dropdown option that is already there.